cosmrs: refactor into submodules
Factors code out of toplevel modules (which map 1:1 to the Go modules)
into type-specific submodules.

This makes it easier to locate the definition of a specific type by
filename, and generally makes it easier to peruse the codebase.
